gitlabci

+++ global-setup.ts
@@ -3,7 +3,9 @@ import { request } from '@playwright/test';
 const database = require('./support/database/functions')
 
 async function globalSetup() {
-    await database.dropDatabase()
+    if (process.env.NODE_ENV != 'production') {
+        await database.dropDatabase()
+    }
 
     const requestContext = await request.newContext();
